fix(ui): refactor for geojson components
