+++ ui/src/components/visualizations/MapLibre/sources/index.ts
@@ -1,0 +1,1 @@
+export * from "./GeoJson";
